Split CPDFXFA_Document apart
[DO NOT COMMIT]

This CL renames IXFA_DocProvider to IXFA_DocEnvironment to better describe the
purpose. Then, CPDFXFA_Document has all of the IXFA_DocEnvironment methods
removed and placed in CPDFXFA_DocEnvironment. The CPDFXFA_Document then
has a CPDFXFA_DocEnvironment.

This splits the code related to the document apart from the XFA callback methods
to work with that document.

Review-Url: https://codereview.chromium.org/2328573002