Use either /RECT or /QuadPoints for annotation coordinates, depending on /AP
On Acrobat, if "/AP" is present on a text markup definition,
the coordinates used to draw the annotation come from "/Rect
values, whereas if "/AP" is not defined, the array defined
in /QuadPoints is used to grab the annotation coordinates from.
PDFium, on the other hand, uses "/Rect" regardless of
presence or absence of "/AP".

CL fixes PDFium to work similarly to Acrobat, in this case.

TEST=testing/resources/pixel/bug_585_*.in
BUG=pdfium:585

Review-Url: https://codereview.chromium.org/2289293005