Assert that dictionary can own the objects it is given.
Upon indirect object holder destruction, all indirect
objects are destroyed -- currently by order of increasing
object number -- but ideally without ordering constraints.

So currently, we can get away with a dictionary pointing
directly at an indirect object with a higher number. It
gets destroyed first, invoking Release() on its subordinates,
which skips destroying them if they are indirect objects. But
we don't want to rely on this artifact of destruction
order.  Should it happen to be reversed, the dictionary
would invoke Release() on freed memory.

Interestingly, CPDF_Array skirts the issue by replacing
any indirect objects it is given with references. Not
clear whether we should do the same thing for dictionaries,
or remove it from arrays.  The technique certainly
complicates understanding ownership.

The one violation found is in the unittest that broke the
previous CL which tried to use unique_ptrs in indirect
object holder.

Review-Url: https://codereview.chromium.org/2353093002