Fix leaks related to the usage of JSE runtime data
Per isolate runtime data should be deleted when the associated
isolate's destructed.

Also, the internal of per isolate runtime data is obscure to the JS
engine. So XFA or this class itself has to be in charge of the memory
management. Use smart pointer for it so that the resource
could be released properly.

BUG=pdfium:242

Review-Url: https://codereview.chromium.org/2354923003