Fixup MSan embeddertests
The embeddertests were closing the document before the formfill environment.
This caused a use-after-free as we try to use the document during formfill
destruction.

This Cl fixes the destruction order in the embedder tests. As well, a few guards
are put in place to keep the system from crashing if the wrong destruction
order is called.

R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/2398063002 .