Fix negative font rendering on Mac by reversing matrix c component.
Notice that this is just making it analoguous to how it works when the
font is positive: the b and d components are reversed. Currently, when
the font is negative, only the a component is reversed. The c one needs
to be reversed as well.

BUG=chromium:653941

Review-Url: https://codereview.chromium.org/2411833002
(cherry picked from commit 4ea4281ad5775686317ef53f48eab39898d13d23)

R=dsinclair@chromium.org
TBR=dsinclair

Review URL: https://codereview.chromium.org/2469443002 .