Rename IFWL_Widget::SetDelegate
The ::SetDelegate method was misleading. The primary use was
SetDelegate(nullptr) which returned the current delegate and didn't actually
set anything. When a value was passed it would set the
|m_pCurDelegate| not the |m_pDelegate|.

This Cl breaks ::SetDelegate into ::GetCurrentDelegate and ::SetCurrentDelegate
to make it clear what is happening and that this does not effect the
|m_pDelegate| variable.

Review-Url: https://codereview.chromium.org/2459423003