Add fpdfppo_embeddertest.cpp.
The lack of coverage of the fpdfppo APIs was noticed while
trying to diagnose another issue.  Adding basic calls to
these APIs then kicked out an assert in XFA, where duplicate
global CFXA_TimeZoneProviders were not expected.  These are
cheap to create except for the global C RTL tzset() call, so
keep track of that and make these on demand.

Review-Url: https://codereview.chromium.org/2488403004