Properly release caches in CPDF_DocRenderData
Currently, only RemoveRef() is being called when releasing. The problem
is that when the object is no longer being used, the map is not updated
accordingly. In the pdf in the bug below, the pointer of a font used in
the second page coincides with the pointer of a font in the first page
(that was released). So the cached values for the font in the first
page are used, resulting in incorrect rendering.

BUG=pdfium:629

Review-Url: https://codereview.chromium.org/2501053002