use skia text
This adds text to the _SUPPORT_SKIA_PATHS_
variant.

The output of the test corpus has more
differences, but no feature changes or
dropouts that I can see.

The text is a bit lighter. Bungeman thinks
this is because of our support of SRGB
color, although the fonts also look different
so maybe there's more to the story.

The output looks compatible with the
variations I see in Adobe Reader and Firefox
on the same platform.

There's probably more text tuning to do.

Additionally, this turns off the caching
feature for this variation. Caching can't
work because the drawing alternates between
Skia and PDFium's native blits, so any
state caching won't know about changes
and drawing performed by the other.

R=dsinclair@chromium.org, bungeman@chromium.org

Review-Url: https://codereview.chromium.org/2520483002