fix negative font size
PDF can specify a negative font size
to change the y-axis for text. Add
support in the Skia driver for that.

The cached form for text and paths
had fallen out of sync with the uncached
code used by skiapaths, so sync that
up so that the mainline skia setting
has those bug fixes as well.

This does not fix the underlying bug in
528103, but gets Skia on parity with
native PDFium text rendering. The bold
text is still drawn in the wrong place.

R=dsinclair@chromium.org
BUG=528103

Review-Url: https://codereview.chromium.org/2528663002