enable skia gradients, fix bugs
Allow shading to use Skia for some gradients;
linear, radial, and coons patch.

The corpus uncovered a few gradient-related bugs
in the Skia driver which are fixed as well.

Fix corpus 11.pdf by ignoring zero area
filled paths. PDFium uses CFX_PathData::GetZeroAreaPath()
to determine if a path that doesn't enclosing anything
should be drawn anyway, and CFX_AggDeviceDriver::GetDriverType()
returns 1 so that the empty path is tweaked so that
it doesn't draw too often.

There are no tests in the corpus that show
the utility of this approach; 11.pdf draws correctly
if the zero area paths are ignored.

Standardize debugging output so that
a separate tool can visualize it.

R=dsinclair@chromium.org, thestig@chromium.org

Review-Url: https://codereview.chromium.org/2530343004