Kill last use of CFX_PtrList.
I'd been waiting on this hoping the code would be refactored out
of existence, but it looks to be used.

Use tests against x.end() in place of a "null position".
This is greatly complicated by the use of "tail position", which
now calcluated by std::prev(x.end()) unless the list is empty.

Review-Url: https://codereview.chromium.org/2592163002