[Merge M56] Add ObservedPtrs to PWL_CREATEPARAM
It's possible for both the provider and attached widget to be destroyed before
the PWL_CREATEPARAM objects which point to them. This causes issues when those
widgets access their attached widget or provider.

This CL wraps the pAttachedWidget and pProvider into ObservedPtrs so we will
know if the underlying pointer has gone away.

In order to merge I also had to add observed ptr copy and assign methods.

BUG=chromium:681351

Change-Id: Ic297753b5d020c5d9401cb3ff4d541b2498346e4
Reviewed-on: https://pdfium-review.googlesource.com/2351
Reviewed-by: Tom Sepez <tsepez@chromium.org>