LibTIFF: fix defines in tiffconf
This CL hardcodes the defines  used in tiffconf so that the values can
be used inside of '#if'. The CL that changed them was:
https://pdfium-review.googlesource.com/c/9410

SIZEOF_VOIDP was a new variable in that CL. This CL uses __LP64__ to
detect whether this value should be set to 8 or to 4.

Bug: chromium:718494
Change-Id: I628d64cb7e2e94c47b8bcc1856abf5949d6578d4
Reviewed-on: https://pdfium-review.googlesource.com/10090
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>