Remove ASSERT that was added to understand what was occuring
thestig provided a test PDF that reproduced the issue that is being
tested for in the ASSERT. I have spent some time going throught the
related code, and the condition in the assert is actually standard
behaviour, so shouldn't be asserted. The following code gracefully
handles the case of more text then requested being returned.

BUG=chromium:763369

Change-Id: I5bc121977169deead52fc5dd2503376b1b62d83f
Reviewed-on: https://pdfium-review.googlesource.com/13750
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>