Consistently apply JS size check
This CL makes the use of CXFA_IsTooBig consistent and universal across
all of the ToJavascript and related methods. Previously this method
was only applied in some calls. It is now being tested as a 
precondition and postcondition on every call and as a post condition 
for the entire translation process.

There are some size checks within methods that potentially generate
large amounts of JS that have been left in.

BUG=chromium:752920

Change-Id: I1a8bfe21e0a17c2e47592fc6017060243674f1bc
Reviewed-on: https://pdfium-review.googlesource.com/16812
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>