Convert CFX_CSSPropertyTable to CFX_CSSData
This converts the existing class naming and references. The ::Entry
struct is converted to ::Property. The GetBy* methods are renamed to
GetPropertyBy*.

Free functions from other parts of the CSS code base are merged into
the class. GetCSSPropertyValueByName becomes ::GetPropertyValueByName
with a return value of ::PropertyValue*. GetCSSLengthUnitByName
becomes ::GetLengthUnitByName with a return value of
::LengthUnit*. GetCSSColorByName becomes ::GetColorByName with a
return value of ::Color*.

BUG=pdfium:998

Change-Id: I5b78c69d0fe53ad714ac3545cfdda8c26b95e4d5
Reviewed-on: https://pdfium-review.googlesource.com/25010
Reviewed-by: dsinclair <dsinclair@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>