[fwl] Restrict setting cursor outside text bounds
This Cl updates CFWL_Edit to keep the cursor from extending pass the end
of the text. If we set the cursor past the end it will trigger the
ASSERT in the text edit engine as the edit engine assumes the caller
does the right thing.

Bug: pdfium:1023
Change-Id: Idfed4b240c13d666a686d0e09b78bcc9a1406794
Reviewed-on: https://pdfium-review.googlesource.com/29650
Commit-Queue: dsinclair <dsinclair@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>