Fix caret moving to start of edit on click after the text.
It now moves to the end of the edit text instead.

In a multiline edit, clicking after the end of a line in the middle
of the text moves the caret to the beginning of the next line. This
is a known issue to be addressed later.

Bug: chromium:832293
Change-Id: I55ca1f7eebf935fbb7d9526f115489bdfbd8ba1c
Reviewed-on: https://pdfium-review.googlesource.com/31012
Reviewed-by: dsinclair <dsinclair@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>