Allow failing to build CFXA_FWLTheme if unable to load fonts
The current implementation of this class potentially does a lot of
work in the constructor. Specifically when getting a calendar font it
might cause the whole font loading pipeline to run. This can fail if
it is unable to load fonts.

Breaking out the font loading part, so that the factory method can
return nullptr if it fails. Additionally adding a guard for the case
the font manager fails to load, which is the root cause of the crash
in the bug.

BUG=chromium:835608

Change-Id: I05b987aaad6f0814907066904331610a7fbb7f70
Reviewed-on: https://pdfium-review.googlesource.com/31330
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>