Add expectations for xfa_node_caption.pdf
This test appears to just be confirming that caption layout and
content can be controlled via JS. Specifically there is a bunch of
listeners for the initialize event that run and setup the page by
inspecting and changing values. Clicking and such does nothing
interesting, so there is no .evt.

BUG=pdfium:1020

Change-Id: I1a24dc6fab058c17362506cce1f184e62dd82500
Reviewed-on: https://pdfium-review.googlesource.com/34491
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>