Fix issue with resolveNodes() not found off global proxy object.
We used to assume that a global proxy object could be distinguished
by it not having two internal fields, but that invariant isn't
correct.  Instead, flag it as such so the block of code at line 126
will check the prototype to find an actual object.

Squeeze some bytes out of the tags while were at it, no reason for
them to be wide. Also remove GetGlobalObjectFromContext() helper, for
transparency into what's really going on in v8. This then shows a
needless retrieval of an object we already have in one case.

Bug: pdfium:1097
Change-Id: Iafc356373166fe5fda76ea7d64193826ee69a6c3
Reviewed-on: https://pdfium-review.googlesource.com/34630
Reviewed-by: dsinclair <dsinclair@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>