Remove explicit calls to FXFT_Clear_Face_External_Stream().
This landed earlier as part of a larger change that got reverted for
unrelated reasons. Now re-land it by itself. See
  https://pdfium-review.googlesource.com/c/pdfium/+/33370

The current code could cause FT to try to free memory allocated by the
caller, with potentially a different allocator. It appears that this may
have been added ages ago as a way to kludge around a leak. If we're
leaking, msan will find it, and we can deal with it properly rather
than trying to force FT to tidy after us.

Change-Id: I20abf4cd35d465de1f435ebe6d94b10c3ff74002
Reviewed-on: https://pdfium-review.googlesource.com/34930
Reviewed-by: dsinclair <dsinclair@chromium.org>
Commit-Queue: dsinclair <dsinclair@chromium.org>