Make empty document always have empty privileges
This landed earlier as part of a larger change that got reverted for
unrelated reasons. Now re-land it by itself. See
  https://pdfium-review.googlesource.com/c/pdfium/+/33370

If we don't have a document, then we can't dynamically tell if it is
an XFA doc, so there's no reason to believe it should have all permissions.
If there is an XFA doc under the covers, then the extension will give
us the value.

Change-Id: I6e007ab659914c39d2e5f3fe7a25d9d13439d358
Reviewed-on: https://pdfium-review.googlesource.com/34932
Commit-Queue: dsinclair <dsinclair@chromium.org>
Reviewed-by: dsinclair <dsinclair@chromium.org>