Fill m_ContentStream field in CPDF_PageObject.
From the comment of CPDF_PageObject::GetContentStream():

"""
Get what content stream the object was parsed from in its page.

This number is the index of the content stream in the "Contents" array,
or 0 if there is a single content stream. If the object is newly
created, -1 is returned.

If the object is spread among more than one content stream, this is
the index of the last one.
"""

Bug: pdfium:1051
Change-Id: I9f7804af4f263dda0422e9542e025e3320ff7c31
Reviewed-on: https://pdfium-review.googlesource.com/34250
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: dsinclair <dsinclair@chromium.org>