Add a way to save a barcode generated bitmap to a .png.
This is useful for debugging and validation purposes. The hashes that
are in BarcodeTest are not necessarily for valid outputs.

This CL refactors the code in embedder_test.png that already does this,
moving it to testing/utils where unit tests can access it too.

Bug: pdfium:1135
Change-Id: I6f1d70a4e133f8f04dbe52646087f99c448e95f8
Reviewed-on: https://pdfium-review.googlesource.com/40152
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>