Optimize rendering of two dimensional barcodes: defer upscale.
Defer upscaling as late as possible so that intermediary data
structures are smaller.

Made a couple of changes along the way to preserve the barcode
correctness and fix some padding issues.

For my example, this is a ~21x improvement in rendering time, down
from ~190ms per barcode to ~9ms.

Bug: 872907, pdfium:1135
Change-Id: If532e0f168f02fea9c31d473f34c0009da4f4612
Reviewed-on: https://pdfium-review.googlesource.com/40010
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>