Make ContentParam::m_Name be a ByteString.
Currently this is an inline fixed-length char[] buffer.  We make a byte
string out of this in many places, and the current implementation still
memcpy's, so the additional costs should be minimal.

Next, we can avoid special-casing names that are longer than the
fixed size.

Change-Id: I980463cbb2325a9d6080bb51a6dfb0dbd1b704b1
Reviewed-on: https://pdfium-review.googlesource.com/42430
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>