Use CropBox instead of ArtBox or TrimBox
FPDFPage_Flatten() used ArtBox as the BBbox of the newly created
XObject, which in some situation, some content of the generated
PDF are not visible.
FPDF_ImportNPagesToOne() uses TrimBox as the BBbox of the newly
created XObject, which in some cases, MediaBox and CropBox are
scaled up, however TrimBox is not, which caused some content of
the newly generated N-upped PDF not visible.
Hence for the above two situations, we have chosen to use mostly
commonly used CropBox.

Bug:409670
Change-Id: Ifb82a6f881d7ce1802cf23c7e8e6f11cc76bf3e9
Reviewed-on: https://pdfium-review.googlesource.com/c/43987
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Shirleen Lou <xlou@chromium.org>