Rename CPDF_Object::Type enumerators with k prefixes
This was originally motivated by Clang's newly enhanced -Wshadow
warning, which complained about BOOLEAN shadowing a BOOLEAN typedef
in winnt.h. The warning has since been changed to not fire on enums
shadowing types, but it was suggested that this kind of naming change
was desirable anyway.

Bug: chromium:895475
Change-Id: I278ce9dbbd88f802fe5bc77fca28b082aa5e684d
Reviewed-on: https://pdfium-review.googlesource.com/c/44330
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>