Stop transfering ownership of |this| as CPWL_Wnd::Create() side-effect
Instead, make it an explicit method called on the parent outside
of create. Rename Create() to Realize() to be sure to catch all
usages, and add the new required call.

Attachment to the parent now takes place a little earlier on in
the life-cycle as a result but should be ok.

Precursor to converting to smart pointers.

Change-Id: I45c459fcd28b5d03c428ce5809d0432506cf4ec6
Reviewed-on: https://pdfium-review.googlesource.com/c/44690
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>