cp.pParentWindow is always NULL in NewPWLWindow()
Because the CreateParams come from one of the GetCreateParam()
overrides, and none of these set that particual field. This
implies that the caller will always own the result, since it
is never handed off by an AddChild() call, which is consistent
with the function return.

Change-Id: I1a9f580f347e911855c695807795541282d43f3f
Reviewed-on: https://pdfium-review.googlesource.com/c/44691
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>