Add CPDF_IndirectObjectHolder::New<T>()
Similar to CPDF_IndirectObjectHolder::NewIndirect<T>(), but returns
an owned object that is capable of interning strings with the indirect
object holder's string pool.

Split off from a forthcoming CL. Eventually, we'd like objects created
in this manner to automatically orphan some objects back to the
indirect object holder under a few cases, and this greatly simplifies
the ability to create the proper objects.

Change-Id: I2b06b98372fa5a510fd424a07b5cb074f8035381
Reviewed-on: https://pdfium-review.googlesource.com/c/45410
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>