Add declarations of missing XFA attributes.
These appear in the (element,attribute) table but not in the
attribute table itself. This is an obstacle to "normalizing"
these tables.

In turn, an XFA_Attribute will no longer fit in a uint8_t, as
caught by our tests, so bump up to int16_t, using a signed
representation for consistency with the other related enums.

Change-Id: I207a81d7a01c6db3419a3b88818825791abbe747
Reviewed-on: https://pdfium-review.googlesource.com/c/46830
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>