Remove fragile script index mechanism.
Replace a index lookup and linear scan with a binary search
over a compound key. Should be able to add/delete elements
easily after this last step.

Move element-attribute information to .inc file. Move the corresponding
element name out of a comment and into the macro itself so that we may
use it as part of that key.


Change-Id: Ib97a7f54c167c3f3323c35a51b336d00f347a689
Reviewed-on: https://pdfium-review.googlesource.com/c/47052
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>