Do not check null return from CFWL_Widget::GetOwnerApp().
Because it can't return null. The underlying member returned
by this method is const, and the constructor immediately derefernces
it to initialize other members, so in turn it can't be null. It is
a good thing, since there are other existing places that omit
the check.

Remove some no-op casts along the way.

Change-Id: I0ecd4bd7ff46d1c074ef90bba0792ee593d74518
Reviewed-on: https://pdfium-review.googlesource.com/c/48313
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>