Split off CFX_RetainablePathData from CFX_PathData.
Because there are places where we use CFX_PathData in a non-refcounted
manner. In turn, update SharedCopyOnWrite so that it can cope with
classes that can only be instantiated via MakeRetain.

This allows us to add a move ctor for CFX_PathData to allow using it
in, say, std::vector<CFX_PathData> without incurring copying costs.

Document the subtlety going on in CFX_RetainablePathData's copy ctor.

Change-Id: Id5a7095558775abf11e4ae901a47f41a6fa5614f
Reviewed-on: https://pdfium-review.googlesource.com/c/49170
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>