Make pAttribute parameter required in LoadImageInfo().
CCodec_ProgressiveDecoder::LoadImageInfo() has only 1 non-test caller
and it passes in a valid pAttribute pointer. Change the call to
LoadImageInfo() from the fuzzer to do the same. Then update the
CodecModuleIface subclasses that fill in the attributes to also have a
required pAttribute parameter.

Change-Id: Ibc824075d858ba86bbd3ee12d8c8e628dd4d4c28
Reviewed-on: https://pdfium-review.googlesource.com/c/49135
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>