Fix FPDF_FILEACCESS lifetime issue with FPDF_LoadCustomDocument().
The FPDF_FILEACCESS that gets passed to FPDF_LoadCustomDocument() does
not need to outlive the returned FPDF_DOCUMENT. See bug for discussion.

This is a partial revert of commit b2ae83e8, along with a test case for
this scenario and a clarification in the public API comments.

BUG=pdfium:1261

Change-Id: Ib1d1b759fe3af98aa4ecc1d8387a259ee4ffe07f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/51570
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>