Fix another CHECK() in CFGAS_StringFormatter::FormatNum()
This was broken in f234dde95a which introduced a length calculation
which was off by one. Additionally, it looks like the calculation
would be wrong since we have the position inside the new string which
is concatenated with the old one at line 1017, so the position needs
adjustment based on the length of the old string.

In turn, this makes an old test actually pass.

Bug: pdfium:1271
Change-Id: I3de697bfc29f7dc5b133e5bc7c08f574ff9eeebd
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/52173
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>