Avoid 0 byte reads in ParseAndAppendCrossRefSubsectionData().
The existing CPDF_Parser::ParseAndAppendCrossRefSubsectionData() code
does not handle reads properly when the read count is a multiple of the
block size. e.g. For a count of 2048 and a block size of 1024, it tries
to read 1024, 1024, and 0 bytes.

Commit cef91f11 added a check that causes 0 byte reads to fail, so the
existing behavior causes ParseAndAppendCrossRefSubsectionData() to fail
altogether. Fix how ParseAndAppendCrossRefSubsectionData() loops to
avoid the 0 byte read.

BUG=chromium:945624

Change-Id: Ie309dd6c3e139720f19962ec2a094512282aa8ed
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/52410
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>