Pass argument by value to CFDF_Document::ParseStream().
Avoids a spurious clang-tidy warning, but is more efficient since
we can move twice and avoid ref-churn. Generally, if a routine is
going to stash away a RetainPtr<>, passing by value and then moving
is a win because the caller might already have an rvalue.

Change-Id: I15bf4ab201d0752ff1ed358ce823629b2156cc81
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/52510
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>