Update IsGdiAvailable to check for Win32k lockdown
If GDI is truly not enabled there should be no guarantee that calling a
GDI method can actually work in any fashion. If gdi32.dll is
delayloaded, for example, calling into it can cause unnecessary loading
and isn't a good practice for something that can't be used anyway. This
change updates IsGdiAvailable to check for Win32k lockdown in the same
way that it is done in the Chromium codebase in
base::win::IsUser32AndGdi32Available().

Bug: pdfium:1278
Change-Id: Ia448e447470a76c65b44bd966b3e7fc108b9524a
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/52570
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>