Rename CXFA_ContainerLayoutItem to CXFA_ViewLayoutItem.
Because the names Content/Container are confusingly similar. The
only sub-class is a CFXA_FFPageView, so the name "view" was chosen.

It is doubly confusing because CFXA_Nodes also have a container
property (as in can contain other nodes) that has no relation to
the layout item containers. That property remains named "container".

It is triply confusing because there is an XFA_Element::ContentArea
node that corresponded to a plain ContainerLayoutItem, not a
ContentLayoutItem.

No functional change. No manual edits.

Change-Id: Ic123599aa10c8656d878e446314e84e9e9a89545
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/52930
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>