Replace dwStatus arguments with specific enums
Because only one of the N bitflags is relevant in a given context,
and it is too easy to pass the wrong values.

Remove dwStatus argument from GetBBox(), it is not used and was
being passed inconsistently by callers. Also replace one boolean
bEllipse in CFXA_FFField with a better enum.

Change-Id: Ibec07f1ce64b5a5965889b8505649d8b271d6fae
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/53055
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>