Fix integer underflow in cfgas_stringformatter.cpp, part 2
A prior CL fixed the possibility of overflow, but there is also a
possibility of underflow if there are several 'E' directives in the
input format string and we have previously negated the exponent.

So discard any accumulated exponent before parsing the exponent again.

Bug: chromium:947188
Change-Id: Ie4236f3b6d955810e0481575a7b73beb458ccd23
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/53350
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>