Use correct layer default configuration
The default layer configuration is contained in the `OCProperties.D`
dictionary. Under certain circumstances, one of the configurations in
`Configs` should be used.

The code was looking for the first config entry with the `Intent` set
to `View` and also specifying the default value of `View`. This means
the first configuration entry that didn't have an `Intent` set was
used, which is wrong. The `Intent` is optional. By removing the default
value, we get the correct behavior.

The test checks that the correct layer configuration, `D`, is used by
making sure the correct image is rendered.

Bug: pdfium:1283

Change-Id: Icac852605e001151cec9cbefec17d06bf93f6e93
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/53370
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>