Move pdfium sources into fpdfsdk
This is necessary to prevent targets like test_support from depending
on pdfium.  test_support depends on pdfium internals, so should not
depend on pdfium when it's a shared library because of ODR violations.

With this change, the dependency on pdfium can be removed (in a followup).

BUG=chromium:960288
R=thestig

Change-Id: I3a65a30b07f3b05bb23e88629546c578983089f0
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/54352
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>