Add tests for FPDF_SetPrintMode().
Set the mode to FPDF_PRINTMODE_POSTSCRIPT2 and
FPDF_PRINTMODE_POSTSCRIPT3, and try to generate some EMFs
that encapsulate PS data.

Add EmbedderTest::GetPostScriptFromEmf(), which is a simplified version
of Chromium code with similar functions, and use it to extract out the
PS data for comparison in tests.

Change-Id: I5cff4ebf0fbac60f63f07b98490a45b8ee3515df
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/55091
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>