Fix CFX_GEModule destruction order.
The platform must outlive the font mgr (which owns the FT library)
which must outlive the font cache. This will enable faces to be
properly ref-counted in a subsequent patch, since the font cache
won't keep faces alive beyond the lifespan of the library.

Change-Id: I448dea06cf38cb62c604c9a71ff89404823f9adc
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/54850
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>