Fix fx_skia_device.cpp so it builds when SHOW_SKIA_PATH is set.
SHOW_SKIA_PATH is a #define that has to be manually set to print out
debugging information, thus the code it guards is normally not compiled.
The code has rotted a bit, so refresh it by adding getters to
CFX_SkiaDeviceDriver and updating the debugging code to use them.

Change-Id: Ic59ad7b846255c376a13aca2a42de9112556e1e3
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/55370
Reviewed-by: Mike Reed <reed@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>