Move CFX_SeekableMultiStream to CPDF_SeekableMultiStream.
Although it implements a fxcrt interface, it does so via fpdfapi/parser
objects, so it must live above fxcrt, since fxcrt shouldn't be allowed
to include fpdfapi. Make it a part of fpdfapi/parser itself under XFA
builds. In turn, this gets rid of a small component
":seekablemultistream" in the middle of the module dependency chart.

Change-Id: Iff47c8329e142359ba921653056470244b911f32
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/55610
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>