Fix layering from fpdfapi/parser to fpdfapi/page and fpdfapi/render.
Almost. There is still a use of the page module in CPDF_Document
that has to move back up into the page layer.

This puts some more burden on CPDF_Document creators to supply
objects from other layers, but that is part of the price of
separation.

Note that test files, although occurring in the same directory as
the code they test, are part of a much higher-level unittests module
which can make use of things the "shipping" code can't. We add some
dependencies to these tests to cover this case.

Change-Id: I185183fcf79b344a78c54600ca021a6765565e82
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/55671
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>