[SkiaPath] Fix issue with missing line.
In bug_1296.pdf, the black line was drawn, but not visible. This was
because it was still using the matrix from the previous operation that
drew the green line. To fix this, reset |SkiaState::m_drawMatrix| to
the identity matrix in FlushDraw() and FlushText().

BUG=pdfium:1296
Change-Id: Ie93f8e1055d10210dfe08f53bc20f48d837360d6
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/54690
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>