Weakly cache CFX_FontMgr::FontDescs.
The intent is keep the FontDesc alive only so long as there at least
one CFX_Face that is still using it. This is similar to the
original manually counted implementation from FX than the current
implementation where we keep the FontDesc around for the lifetime
of the FontMgr.

- Avoid some circularity between CFX_Face and CFX_FontMgr::FontDesc
  by only telling CFX_Face that it has a retainable.

Change-Id: If8b441b9b1c2aa81892cb4cb62762a44a322dd0c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/56652
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>