API to return text range for weblinks
PDF text is parsed to find links using regex matching. These are not
links which are present as annotations in the PDF but are still
considered links (categorized as Auto-Detected links).
While parsing the links PDFium identifies the start char index and char
count of the detected link.

This CL exposes a public method to get start char index and char count
of such auto detected links. This is required for accessibility to
generate the relative mapping of links and text.

This CL adds a unit test to cover both cases: asking the text range for
a valid link and an invalid link.

Bug: pdfium:1334
Change-Id: I6d98b35eb9c6e7fcf4d800776b9129588fce0e09
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/56810
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>