Remove const from public fpdfview types.
They contain (or are aliases for) parser objects, which are now
refcounted, and order to manipulate the internal refcounts, non-const
objects must be used. This avoids a number of const casts throughout
the codebase. Regrettably, there is some loss of const as a result.

This CL must happen first before we can use RetainPtr in a
number of these classes.

Change-Id: I0e894b56373ad35c68daf4499255be582699378d
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/57853
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>