CJX_Object::SetContent() should only update value nodes.
This method operates in two passes. In the first part, a set
of value nodes children is collected, but in the second pass,
all children are traversed regardless of whether they are
value nodes. There's no reason to believe these are the same
when the node has non-value node children.

- Fix possibly overly-strong assert in CXFA_Node. In turn, this
  causes one test to leave a node unmodified rather than CHECK().

Bug: chromium:985781
Change-Id: Idd8ae5d8fb5f07ae01487db589036bdab3a99ada
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/57932
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>