Mark fxcrt::RetainPtr<>::m_nRefCount as mutable.
Similar to base, this will allow us to have RetainPtrs to objects
that are otherwise const apart from the ref counts. In turn, make
Retain() and Release() const, which leads to the odd situation where
we delete a const |this| (but legal in C++ for a variety of reasons).

Change-Id: I975628c5c60e6337afd496ade2371a4c1b69e84a
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/58191
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>