Make CPDF_Font a retainable type.
In turn, this makes CPDF_CountedObject<> unused, so remove it.

We have to explicitly break some cycles between fonts and forms,
so introduce a WillBeDestroyed() method.

This also introduces the requirement that callers of
FPDFText_LoadStandardFont() release the resource via a call
to FPDFFont_Close().


Change-Id: I27451a24e55424e7f784a1af962036269295aed2
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/57470
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>