Always pass NULL |hWidget| in FFI_PopupMenu().
This is an API object of type FPDF_WIDGET, and there aren't any
public interfaces for the caller to do anything with that kind
of an object. About all an embedder could do would be to check
if it is the same as in a previous call, but the underlying
object is deep in XFA and Observable, and may come and go at will.

Handing out a pointer to such an XFA object feels dangerous. So
make it NULL despite the public API change.

Change-Id: I6a97b40b4d58b025d0b7c1928d67be4f22be6e40
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/59311
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>