Convert bug_512557.in to bug_512557.pdf.
The .in files should be text files with LF line endings, so they should
not contain binary data. For bug_512557.in, the trigger for the bug
requires the image in the PDF to have /Filter /RunLengthDecode, and the
usual workaround of also ASCIIHex encoding the data does not work. Given
this, just convert the file to PDF.

Do the same for bug_304.in, which may be sensitive to the presence of
additional filters.

Bug: pdfium:1302
Change-Id: I6e0179e1968ea47e48b8c2cf3b1b5027512f2c4d
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/60572
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>