Fix the TrueType font not rendered correctly issue.
This CL removes the extra check on whether no glyph index is found for
a TrueType font character in CPDF_SimpleFont::GlyphFromCharCode(), so
that whenever CPDF_SimpleFont::GlyphFromCharCode() returns -1, it
indicates a guaranteed failure and we can handle the TrueType font
later on in CPDF_CharPosList::CPDF_CharPosList().

If a TrueType font fails to find glyph index, i.e. when
CPDF_SimpleFont::GlyphFromCharCode() returns 0, we should first try to
find the glyph index from the fallback font. If that fails, switch back
to use the original font.

Bug: pdfium:1388
Change-Id: I8b1935a094f76408e4678723d566914c2bbf0f42
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/60091
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>