Cover some JS public methods that do not "commit".
The path taken sometimes depends on the nature of the event which
triggered the JS execution and the callback, so test the public
method callbacks under different types of triggering events. In
particular, pick an event that does not set the "will commit" bit
in the event recording structure. To trigger this, we need to
send a keystroke to the appropriate field.

Change-Id: I9aabefc1e391c5a89acddd6b3a96fe6337f142d7
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/62212
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>