Fix LoadLinearizedAllCrossRefV4() to work properly.
In CPDF_Parser, for linearized PDFs, LoadLinearizedAllCrossRefV4() has
to read the first-page cross-reference table, as well as the main xref
table. It has to read the associated trailers for those xref tables as
well, and also merge them.

Currently, LoadLinearizedAllCrossRefV4() does not do this correctly. It
loads the main trailer, and merges it with the first-page trailer before
ever reading from the first-page trailer. Therefore, when it does read
/XRefStm for the first time, it will always fail and get values of 0.

Fix this by reading from the first-page trailer first, and then merging
the two trailers. Add comments in the function to explain what is
happening.

Bug: pdfium:1335
Change-Id: Iab5ac79d8e0c25c72519075179b001fc19cece51
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/62551
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>