Add utility for filter-encoding PDF streams
We frequently need to convert binary into one of the /ASCII*Decode
encodings (ideally with /FlateDecode compression) for embedding in .in
files. This process is a bit of a pain in the neck currently, since you
need several separate utilities and steps to do the conversions.

This tool will do it in one step (given a source of binary bytes), and
format it nicely for pasting directly into a .in file.

Change-Id: I4545eabca2dd45748d0cdcedc44a3984131c394c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/63350
Commit-Queue: K Moon <kmoon@chromium.org>
Reviewed-by: Lei Zhang <thestig@google.com>