Avoid regenerating revision 5 and 6 encryption dictionaries.
Unlike revision 2 and 3, revision 5 and 6 encryption is not tied to the
document ID in the trailer. Thus regenerating the encryption dictionary
when the ID changes is completely unnecessary. Avoid doing this, and
allow more CPDFSecurityHandlerEmbedderTests to test opening save copies
with both the owner and user passwords.

Change-Id: I074a9b6e03bcaa39c8fb18eed6487454bdc5bcd1
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/63933
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>