Only hash pixel data in EmbedderTest::HashBitmap().
Add a new EmbedderTest::HashBitmap() implementation that only hashes the
pixels in a given bitmap, and ignores the data in the gaps at the end of
rows. Per prior discussion on [1], the data in the gaps can in theory be
uninitialized. This does not most hashes, as most bitmaps in embedder
tests do not have gaps.

[1] https://pdfium-review.googlesource.com/64290

Change-Id: I24bd143f6b33c7ef86e26dd09d1cb874dea1675f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/64331
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>