[SkiaPaths] Tiling pattern with overlapping paths causes unwanted grid.
When a colored tiling pattern's opacity value is not 1.0, the section
of overlapped patterns will have a different shade of color.
bug_1286.in is not rendered correctly because the patterns' paths
(borders) are overlapping and SkiaPaths chose to use
kDifference_SkPathOp to subtract the latter path from the previous one,
which result in the overlapping paths to have a lighter color than the
pattern itself and displayed as a grid. Need to use kUnion_SkPathOp to
union the overlapping paths and skip stroking the path if the path's
fill color is the same as the stroke color, so that the overlapping
part of the unionized paths won't show in a darker shade.

Bug: pdfium:1342
Change-Id: Ideb665ae42bd077f6fe3be364a056286a520a1e6
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/60230
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Hui Yingst <nigi@chromium.org>