Remove |offsets| parameter from CJPX_Decoder::Decode().
The decoder does not need this level of flexibility. Replace it with
a |swap_rgb| parameter and calculate the offsets internally in a
foolproof manner.

Also rename a local variable in one caller to better describe what it
represents and to avoid Hungarian notation.

Change-Id: I6948517dd4af03d942be08e3e597c1c667e71396
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/65791
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>