Add a null check on return value of CPDF_StructTree::AddPageNode().
This CL Adds a null check for |pParentElement| before usage inside
AddPageNode(). This ensures that the code handles the scenario when
AddPageNode() returns nullptr if depth is greater than
kStructTreeMaxRecursion(= 32). This CL also includes a test tagged PDF
with deeply nested tags(nested depth > 33). The file is used to validate
if struct tree is loaded for this file.

Bug: pdfium:1480
Change-Id: I66c561d6708f8d89a0616eb7adef8459ae952662
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/66110
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Mansi Awasthi <maawas@microsoft.com>