Remove code after abort() for -Wunreachable-code.
The code was added in https://codereview.chromium.org/1128043009/

It was then later ifdef'd out on Windows to enable its unreachable
code warning c4702.

Now that we want to turn this warning on with other compilers,
remove the code completely.

abort() is marked noreturn in headers. It's possible to
install a SIGABRT handler, but it's not allowed to return,
so it can't be used to make abort() return -- see `man abort`.

Bug: chromium:346399
Change-Id: I05b01a8424a66c6609f16af47e40698e9f7e376e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/67410
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>