Fix rendering issue for PDFs with incorrect first-page hint data.
For a linearized PDF, sometimes the hint table can contain incorrect
data. For example, The test file in crbug.com/1012237 has a hint table
which points the first page to the wrong object, and set the wrong
object number into |m_PageList[0]|. This results in fetching a
dictionary of a non-page-type object for the first page and rendering
the first page blank.

To fix the issue, this CL creates CPDF_Document::IsValidPageObject() to
validate if an object's type is "Page", and add this validation in
CPDF_Document::LoadPages(). If the check fails, don't set this wrong
object number to |m_PageList[0]|. Instead, traverse the page tree to
update |m_PageList|.

Bug: chromium:1012237
Change-Id: I1e8745a48ed0cabe5478e9e8e1dda51fa0a8d53f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/67490
Commit-Queue: Hui Yingst <nigi@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>