Change for loops to use const refs where possible.
LSC: Change for loops to use const refs, per performance-for-range-copy
clang-tidy.

This CL optimizes C++11 range-based for loops where the variable is
copied in each iteration but it would suffice to obtain it by const
reference. This is only applied to loop variables of types that are
expensive to copy which means they are not trivially copyable or have
non-trivial copy constructor or destructor.

To ensure that it is safe to replace the copy with a const reference
the following heuristic is employed:
  The loop variable is const qualified.
  The loop variable is not const, but only const methods or operators
    are invoked on it, or it is used as const reference or value
    argument in constructors or function calls.

PiperOrigin-RevId: 304179354
Change-Id: I54a9409ba8fd97ed4bbdc0a930277761c27cc04c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/68110
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>