Fix insertions into empty name trees.
When a name tree is empty, CPDF_NameTree::AddValueAndName() tries to
find a leaf node to insert into and fails. Instead, just insert into the
root node of the name tree. Add a unit test for this case.

This fixes FPDFDoc_AddAttachment() for the case where the EmbeddedFiles
name tree is empty. Fix the known to be broken
AddAttachmentsToFileWithNoAttachments embedder test case and do further
testing.

Bug: pdfium:1502
Change-Id: Iab189ce612070e0ef268ac6e01a1cbcce96c8d2b
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/67990
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>