Add API for removing an inklist for an ink annotation.
There is no API to remove/edit an inklist for an ink annotation.
With this API, user can edit an inklist by removing it and adding it
again with edited ink points.

This CL adds API and relevant tests for removing an inklist for an
ink annotation.

Bug: pdfium:1512
Change-Id: I62fc232928ea720c8cef36adfb81fa20c0b10773
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/68613
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Shikha Walia <shwali@microsoft.com>